+++ react-vite/src/components/Forms/LoginFormPage/LoginFormPage.jsx
@@ -1,15 +1,18 @@
 import { useState } from "react";
-import { thunkLogin } from "../../redux/session";
-import { useDispatch } from "react-redux";
-import { useModal } from "../../context/Modal";
+import { thunkLogin } from "../../../redux/session";
+import { useDispatch, useSelector } from "react-redux";
+import { Navigate, useNavigate } from "react-router-dom";
 import "./LoginForm.css";
 
-function LoginFormModal() {
+function LoginFormPage() {
+  const navigate = useNavigate();
   const dispatch = useDispatch();
+  const sessionUser = useSelector((state) => state.session.user);
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const [errors, setErrors] = useState({});
-  const { closeModal } = useModal();
+
+  if (sessionUser) return <Navigate to="/" replace={true} />;
 
   const handleSubmit = async (e) => {
     e.preventDefault();
@@ -24,13 +27,15 @@ function LoginFormModal() {
     if (serverResponse) {
       setErrors(serverResponse);
     } else {
-      closeModal();
+      navigate("/");
     }
   };
 
   return (
     <>
       <h1>Log In</h1>
+      {errors.length > 0 &&
+        errors.map((message) => <p key={message}>{message}</p>)}
       <form onSubmit={handleSubmit}>
         <label>
           Email
@@ -58,4 +63,4 @@ function LoginFormModal() {
   );
 }
 
-export default LoginFormModal;
+export default LoginFormPage;
