+++ react-vite/src/components/Forms/SignupFormPage/SignupFormPage.jsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { Navigate, useNavigate } from "react-router-dom";
-import { thunkSignup } from "../../redux/session";
+import { thunkSignup } from "../../../redux/session";
 
 function SignupFormPage() {
   const dispatch = useDispatch();
