+++ react-vite/src/components/Modals/SignupFormModal/SignupFormModal.jsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { useDispatch } from "react-redux";
-import { useModal } from "../../context/Modal";
-import { thunkSignup } from "../../redux/session";
+import { useModal } from "../../../context/Modal";
+import { thunkSignup } from "../../../redux/session";
 import "./SignupForm.css";
 
 function SignupFormModal() {
